rename package and module based on kevinb9n's proposal
https://github.com/jspecify/jspecify/issues/1#issuecomment-620826847
