+++ src/main/java/module-info.java
@@ -1,3 +1,3 @@
-module org.jspecify.experimentaldonotuse {
-  exports org.jspecify.experimentaldonotuse;
+module org.jspecify.annotations {
+  exports org.jspecify.annotations;
 }
