+++ src/main/java/org/jspecify/annotations/Nullable.java
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jspecify.experimentaldonotuse;
+package org.jspecify.annotations;
 
-import static java.lang.annotation.ElementType.PACKAGE;
-import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.ElementType.TYPE_USE;
 import static java.lang.annotation.RetentionPolicy.CLASS;
 
 import java.lang.annotation.Documented;
@@ -24,6 +23,6 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
 @Documented
-@Target({TYPE, PACKAGE})
+@Target(TYPE_USE)
 @Retention(CLASS)
-public @interface DefaultNonNull {}
+public @interface Nullable {}
